requirements.txt
Django==5.1.1
django-weasyprint==2.3.0
